Improve exceptions in test fixture
Replace RuntimeException with LogicException, because if you call a
getter before setting a value, you're not using the class right. This is
a developer mistake, not a runtime anomaly.

See https://moxio.com/blog/best-practices-for-php-exception-handling/

This change and the original introduction of the null checks show the
futility of FreezableValueObject in modern PHP, where we're better off
with value objects that have readonly, non-nullable constructor
parameters. Right now, FreezableValueObject also tries to implement the
builder pattern by allowing for null values, but at the cost of having
to add null checks to every getter, which defies the purpose of
`assertNotNullFields`